# openapi/agentteams-public.yaml
openapi: 3.1.0
info:
  title: AgentTeams Public API
  version: v1.0
  description: >-
    Stable public subset for Project and Task management. Internal control-plane,
    Matrix AppService, and Kubernetes APIs are intentionally excluded.
servers:
  - url: /
security:
  - BearerAuth: []
paths:
...
  /api/v1/tasks:
    get:
      operationId: listTasks
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Direction'
        - name: phase
          in: query
          schema: { $ref: '#/components/schemas/TaskPhase' }
        - name: teamId
          in: query
          schema: { type: string, format: uuid }
        - name: workerId
          in: query
          schema: { type: string, format: uuid }
        - name: actor
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Task page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      operationId: createTask
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaskRequest' }
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/tasks/{taskId}/runs/{runId}/progress:
    get:
      operationId: getTaskProgress
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/RunId'
        - name: phase
          in: query
          schema: { type: string, default: EXECUTION }
      responses:
        '200':
          description: Task progress snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskProgressSnapshot' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/tasks/{taskId}/runs/{runId}/result:
    get:
      operationId: getTaskResult
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/Visibility'
      responses:
        '200':
          description: Final task result manifest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResultManifest' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/tasks/{taskId}/runs/{runId}/process-events:
    get:
      operationId: listTaskProcessEvents
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/RunId'
        - name: after
          in: query
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - $ref: '#/components/parameters/Visibility'
      responses:
        '200':
          description: Ordered, replayable task process events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaskProcessEvent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/tasks/{taskId}/runs/{runId}/process-events/stream:
    get:
      operationId: streamTaskProcessEvents
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/LastEventId'
        - $ref: '#/components/parameters/Visibility'
      responses:
        '200':
          description: Server-sent event stream with stable sequence IDs
          content:
            text/event-stream:
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Webhook subscriptions
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/WebhookSubscription' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      operationId: createWebhook
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWebhookRequest' }
      responses:
        '201':
          description: Created Webhook subscription
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookSubscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 512 }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Sort:
      name: sort
      in: query
      required: false
      schema: { type: string, enum: [updatedAt, createdAt] }
    Direction:
      name: direction
      in: query
      required: false
      schema: { type: string, enum: [asc, desc], default: desc }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    OrganizationId:
      name: organizationId
      in: query
      required: true
      schema: { type: string, minLength: 1 }
    TenantId:
      name: tenantId
      in: query
      required: true
      schema: { type: string, minLength: 1 }
    ProjectId:
      name: projectId
      in: query
      required: false
      schema: { type: string, minLength: 1 }
    Visibility:
      name: visibility
      in: query
      required: false
      schema: { type: string, enum: [REQUESTER], default: REQUESTER }
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      schema: { type: integer, format: int64, minimum: 0 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Resource is outside the caller scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Version or idempotency conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    RateLimited:
      description: Retryable rate limit
      headers:
        Retry-After:
          schema: { type: integer, minimum: 0 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServerError:
      description: Server or dependency failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    CursorPage:
      type: object
      required: [items, hasMore, serverTime]
      properties:
        items: { type: array, items: {} }
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
        serverTime: { type: string, format: date-time }
...
    TaskPage:
      allOf:
        - $ref: '#/components/schemas/CursorPage'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/TaskListItem' }
    ApiError:
      type: object
      required: [code, message, correlationId, details]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string }
        details:
          type: object
          additionalProperties: true
...
    CreateTaskRequest:
      type: object
      required: [title, description, spec]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: string }
        spec: { type: object, additionalProperties: true }
        actor: { type: string }
        source: { type: string }
    Task:
      type: object
      required: [id, title, description, phase, priority, createdAt, updatedAt, version]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        phase: { $ref: '#/components/schemas/TaskPhase' }
        priority: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, format: int64, minimum: 0 }
...
    CreateWebhookRequest:
      type: object
      required: [organizationId, tenantId, projectId, endpoint, secretRef, eventTypes]
      additionalProperties: false
      properties:
        organizationId: { type: string, minLength: 1 }
        tenantId: { type: string, minLength: 1 }
        projectId: { type: string, minLength: 1 }
        endpoint: { type: string, format: uri, pattern: '^https://' }
        secretRef: { type: string, minLength: 1, maxLength: 255 }
        eventTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [task.created, task.updated, task.completed, task.failed, task.process, task.result]
    WebhookSubscription:
      type: object
      required: [id, organizationId, tenantId, projectId, endpoint, eventTypes, enabled, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string }
        tenantId: { type: string }
        projectId: { type: string }
        endpoint: { type: string, format: uri }
        eventTypes: { type: array, items: { type: string } }
        enabled: { type: boolean }
        version: { type: integer, format: int64, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskListItem:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            tenantId: { type: string }
            projectId: { type: string }
            team: { type: string }
            actor: { type: string }
            source: { type: string }
            teamId: { type: string, format: uuid }
            workerId: { type: string, format: uuid }
...
    TaskProgressSnapshot:
      type: object
      required: [phase, completed, total, progress, waitingReason]
      properties:
        phase: { type: string }
        completed: { type: integer, format: int64, minimum: 0 }
        total: { type: integer, format: int64, minimum: 0 }
        progress: { type: integer, minimum: 0, maximum: 100 }
        waitingReason: { type: string }
    TaskProcessEvent:
      type: object
      required: [eventId, taskId, runId, sequence, eventType, visibility, occurredAt, correlationId]
      properties:
        eventId: { type: string, format: uuid }
        taskId: { type: string, format: uuid }
        runId: { type: string, format: uuid }
        sequence: { type: integer, format: int64, minimum: 0 }
        eventType: { type: string }
        visibility: { type: string, enum: [REQUESTER] }
        occurredAt: { type: string, format: date-time }
        correlationId: { type: string }
        payload: { type: string }
        payloadRef: { type: string }
    TaskResultManifest:
      type: object
      required: [taskId, runId, status, summary, artifacts]
      properties:
        taskId: { type: string, format: uuid }
        runId: { type: string, format: uuid }
        status: { type: string }
        summary: { type: string }
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/TaskArtifactMetadata' }
    TaskArtifactMetadata:
      type: object
      required: [name, storageRef, contentType, sizeBytes, sha256, version, stage, visibility]
      properties:
        name: { type: string }
        storageRef: { type: string }
        contentType: { type: string }
        sizeBytes: { type: integer, format: int64, minimum: 0 }
        sha256: { type: string }
        version: { type: integer, format: int64, minimum: 0 }
        stage: { type: string }
        visibility: { type: string, enum: [REQUESTER] }